pyproject.toml
[tool.poetry]
name = "rkl-discord-bot"
version = "0.1.0"
description = ""
authors = ["0xCold"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
pandas = "^1.4.1"
numpy = "^1.22.3"
Pillow = "^9.0.1"
boto3 = "^1.21.28"
google = "^3.0.0"
google-api-python-client = "^2.42.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
